backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
python-jose
opencv-python
mediapipe